hotfix 4

+++ gemini/gemini_core/gemini_master.py
@@ -139,8 +139,8 @@ class Gemini:
         begin_price = self.data.iloc[0]['open']
         final_price = self.data.iloc[-1]['close']
 
-        shares = int(float(self.account.initial_capital)) / int(float(self.data.iloc[0]['close']))
-        self.data['base_equity'] = [price * shares for price in
+        shares = float(self.account.initial_capital) / float(self.data.iloc[0]['close'])
+        self.data['base_equity'] = [float(price) * shares for price in
                                     self.data['close']]
         self.data['equity'] = [e for _, e in self.account.equity]
